Refactor project structure and update includes
In this commit, the project structure has been heavily modified to improve clarity. Specific changes include relocating numerous files from 'include' to 'src' to better reflect their roles. Includes were updated to point to the correct locations and to use double-quote syntax instead of angle-brackets. Some unnecessary include directories were removed from CMakeLists.txt and paths to GLFW library files were updated.
