+++ client/src/main.cpp
@@ -8,9 +8,12 @@ class Game : public Ayla::Application{
 public:
     Game() = default;
     ~Game() = default;
+
+    
 };
 
 
 Ayla::Application* Ayla::CreateApplication() {
+
     return new Game();
 }
