+++ src/Ayla/Core/Application.h
@@ -2,9 +2,10 @@
 // Created by sparky on 4/16/2024.
 //
 #pragma once
-#include <Ayla/Core.h>
-#include <Ayla/WindowSystem/Window.h>
-#include <Ayla/WindowSystem/GenericWindow.h>
+#include "Ayla/Core/Core.h"
+#include "Ayla/WindowSystem/Window.h"
+
+
 namespace Ayla {
 
      class AYLA_API Application {
