+++ src/Ayla/EntryPoint/WindowLoop.cpp
@@ -1,14 +1,14 @@
 //
 // Created by Andrew Fagan on 4/16/24.
 //
-#include "Ayla/EntryPoint/WindowLoop.h"
+#include "WindowLoop.h"
 
 
 #include <iostream>
 //#include <SDL.h>
 #include <cstdlib>
 
-#include <Ayla/WindowSystem/GenericWindow.h>
+#include "../WindowSystem/GenericWindow.h"
 
 namespace Ayla::EntryPoint {
