+++ src/Ayla/aypch.h
@@ -25,7 +25,7 @@
 #include <unordered_set>
 
 // Other Stuff
-#include <Ayla/Core.h>
+#include "Ayla/Core/Core.h"
 
 #ifdef AYLA_PLATFORM_WINDOWS
     #include <Windows.h>
